requirements.txt
APScheduler==3.10.4
greenlet==3.0.3
psutil==5.9.8
psycopg2-binary==2.9.9
pyaes==1.6.1
pymediainfo==6.1.0
pyrotgfork==2.1.30
PySocks==1.7.1
python-dotenv==1.0.1
pytz==2024.1
six==1.16.0
SQLAlchemy==2.0.30
TgCrypto==1.2.5
typing_extensions==4.11.0
tzlocal==5.2
uvloop==0.19.0
flask
requests==2.31.0
